+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,23 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Raise InvalidArgumentValueError for azure cni + pod_cidr without overlay.
+
+0.5.125
++++++++
+* Update the minimum required cli core version to `2.44.0`.
+* Support for data collection settings to the AKS Monitoring addon
+* Add `--data-collection-settings` option in aks create and aks enable-addons
+
+0.5.124
++++++++
+* Update command group `az aks maintenanceconfiguration` to support the creation of dedicated maintenance configurations:
+  * *aksManagedAutoUpgradeSchedule* for scheduled cluster auto-upgrade
+  * *aksManagedNodeOSUpgradeSchedule* for scheduled node os auto-upgrade
+
+0.5.123
++++++++
+* Add the KataMshvVmIsolation option to --workload-runtime.
 
 0.5.122
 +++++++